fuzzqr/QRCodeGenerator/requirements.txt
qrcode
Pillow
argparse
base45
flynn
cose
PyPDF2 
pyzbar
pypng
pyqrcode